celery configurado con docker, ya se pueden hacer escenarios de pruebas por el momento a la carga y descarga de archivos
