+++ .gitignore
@@ -150,5 +150,6 @@ files/static/*
 !files/static/.gitkeep
 modelos/__pycache__
 vistas/__pycache__
+application/__pycache__
 .DS_Store
 files/.DS_Store
