+++ helper.py
@@ -1,13 +1,6 @@
 import os
 import hashlib
-import zipfile
-from os.path import basename
-from modelos import db, Tarea, NewFormatEnum, StatusEnum
-
-STATIC_FOLDER = os.path.join(os.getcwd(), 'files/static/')
-MEDIA_FOLDER = os.path.join(os.getcwd(), 'files/media/')
-
-ALLOWED_EXTENSIONS = {'txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif', 'json'}
+from config import ALLOWED_EXTENSIONS
 
 def encrypt(text):
     text = text.encode('utf-8')
@@ -23,25 +16,3 @@ def get_file_path(filename, folder):
 
 def filepath(filename, folder):
     return os.path.join(folder, filename)
-
-def compress_file(tarea_id):
-    tarea = Tarea.query.get(tarea_id)
-    filename = tarea.fileName
-    filename_no_ext = filename.split('.')[0]
-    ext = tarea.newFormat
-    if ext == NewFormatEnum.ZIP:
-        zip_file(filename, filename_no_ext, ext)
-        update_task_state(tarea, StatusEnum.processed)
-
-
-def zip_file(filename, filename_no_ext, new_ext):
-    original_filepath = filepath(filename, MEDIA_FOLDER)
-    new_filename = f"{filename_no_ext}.{new_ext}"
-    new_filepath = get_file_path(new_filename, STATIC_FOLDER)
-    with zipfile.ZipFile(new_filepath, 'w') as f:
-        f.write(original_filepath, basename(original_filepath))
-
-def update_task_state(tarea, state):
-    tarea.status = state
-    db.session.add(tarea)
-    db.session.commit()
