+++ requirements.txt
@@ -1,3 +1,4 @@
+celery==5.2.7
 flask==2.2.3
 flask_cors==3.0.10
 flask_restful==0.3.9
